rag/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
langchain==0.0.340
langchain-community==0.0.4
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.25.2
python-docx==1.0.1
PyPDF2==3.0.1
openpyxl==3.1.2
python-pptx==0.6.22
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
aiofiles
pydantic==2.5.0
httpx==0.25.2
huggingface_hub==0.14.1
python-dotenv==1.0.0
chromadb
